spread some things out, better organized
